// packages/sunshine-admin/src/components/Table/src/types/table.ts
export interface TableActionProps {
  actions?: ActionItem[];
  dropDownActions?: ActionItem[];
  divider?: boolean;
  outside?: boolean;
  stopButtonPropagation?: boolean;
  more?: ReactNode;
}